Development

+++ cozmo_fsm/opengl.py
@@ -9,11 +9,15 @@ from OpenGL.GLU import *
 from threading import Thread  # for backgrounding window
 
 INIT_DONE = False
+MAIN_LOOP_LAUNCHED = False
+
+# Maintain a registry of display functions for our windows
+WINDOW_REGISTRY = []
 
 def init():
     global INIT_DONE
     if not INIT_DONE:
-        print('opengl init calling glutInit()')
+        INIT_DONE = True
         glutInit()
         glutInitDisplayMode(GLUT_RGBA | GLUT_DOUBLE | GLUT_DEPTH)
 
@@ -22,15 +26,26 @@ def init():
 
         # Killing window should not directly kill main program
         glutSetOption(GLUT_ACTION_ON_WINDOW_CLOSE, GLUT_ACTION_CONTINUE_EXECUTION)
-        INIT_DONE = True
 
 def create_window(name,size=(500,500)):
+    global WINDOW_REGISTRY
     init()
     glutInitWindowSize(*size)
     w = glutCreateWindow(name)
+    WINDOW_REGISTRY.append(w)
     return w
 
+def idle():
+    for window in WINDOW_REGISTRY:
+        glutSetWindow(window)
+        glutPostRedisplay()
+
 def launch_main_loop():
+    if not INIT_DONE: return   # no windows were created
+    global MAIN_LOOP_LAUNCHED
+    if MAIN_LOOP_LAUNCHED: return
+    MAIN_LOOP_LAUNCHED = True
+    glutIdleFunc(idle)
     thread = Thread(target=glutMainLoop)
     thread.daemon = True #ending fg program will kill bg program
     thread.start()
